regulation_search/website/api/requirements.txt
azure-functions
azure-ai-projects==1.0.0b3
azure-identity
typing-extensions>=4.0.0
